Hacktoberfest Migration
- Created a `hacktoberfest` super-command to all commands
- Moved util functions into `_utils.py`
- Moved commands into `_cog.py` and renamed as appropriate
- Updated `__init__.py` so that it loads the new `Hacktoberfest` cog
